# devintel-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevIntel AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/repos/{repository_id}/search": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Search Repository",
        "description": "Semantic search across indexed code in a repository.",
        "operationId": "search_repository_api_v1_repos__repository_id__search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Repository Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "title": "Q"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/github": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "List Github Repositories",
        "description": "Fetch repositories from the user's GitHub account.",
        "operationId": "list_github_repositories_api_v1_repos_github_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "List Repositories",
        "description": "List repositories (personal).",
        "operationId": "list_repositories_api_v1_repos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Create Repository",
        "description": "Add a repository.",
        "operationId": "create_repository_api_v1_repos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/index": {
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Index Repository",
        "description": "Trigger repository indexing.",
        "operationId": "index_repository_api_v1_repos_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryIndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryIndexResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/repos/{repository_id}/pulls": {
      "get": {
        "tags": [
          "PR Review"
        ],
        "summary": "List Pull Requests",
        "description": "List pull requests for a repository from GitHub.",
        "operationId": "list_pull_requests_api_v1_repos__repository_id__pulls_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "open",
              "title": "State"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullRequestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repository_id}/pulls/{pr_number}/diff": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get Pull Request Diff",
        "description": "Get the unified diff for a pull request.\n\nReturns raw patch text so the frontend can render a syntax-highlighted diff.",
        "operationId": "get_pull_request_diff_api_v1_repos__repository_id__pulls__pr_number__diff_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Repository Id"
            }
          },
          {
            "name": "pr_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pr Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repository_id}/status": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get Repository Status",
        "description": "Get repository indexing status (lightweight polling endpoint).",
        "operationId": "get_repository_status_api_v1_repos__repository_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repository_id}": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get Repository",
        "description": "Get a repository by ID.",
        "operationId": "get_repository_api_v1_repos__repository_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Repositories"
        ],
        "summary": "Delete Repository",
        "description": "Delete a repository.",
        "operationId": "delete_repository_api_v1_repos__repository_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndexingStatusResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "indexed_status": {
            "type": "string",
            "title": "Indexed Status"
          },
          "indexing_progress": {
            "type": "integer",
            "title": "Indexing Progress"
          },
          "indexing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexing Error"
          },
          "last_indexed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed At"
          },
          "last_indexed_commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed Commit Sha"
          },
          "indexing_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexing Mode"
          }
        },
        "type": "object",
        "required": [
          "id",
          "indexed_status",
          "indexing_progress",
          "indexing_error",
          "last_indexed_at",
          "last_indexed_commit_sha",
          "indexing_mode"
        ],
        "title": "IndexingStatusResponse"
      },
...
      "PullRequestListResponse": {
        "properties": {
          "pulls": {
            "items": {
              "$ref": "#/components/schemas/PullRequestResponse"
            },
            "type": "array",
            "title": "Pulls"
          },
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id"
          }
        },
        "type": "object",
        "required": [
          "pulls",
          "repository_id"
        ],
        "title": "PullRequestListResponse",
        "description": "Pull request list response schema."
      },
...
      "RepositoryCreate": {
        "properties": {
          "repo_name": {
            "type": "string",
            "title": "Repo Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "stars": {
            "type": "integer",
            "title": "Stars",
            "default": 0
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch",
            "default": "main"
          }
        },
        "type": "object",
        "required": [
          "repo_name",
          "full_name",
          "url"
        ],
        "title": "RepositoryCreate"
      },
      "RepositoryIndexRequest": {
        "properties": {
          "repository_id": {
            "type": "integer",
            "title": "Repository Id"
          }
        },
        "type": "object",
        "required": [
          "repository_id"
        ],
        "title": "RepositoryIndexRequest"
      },
      "RepositoryIndexResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "repository_id": {
            "type": "integer",
            "title": "Repository Id"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "message",
          "repository_id"
        ],
        "title": "RepositoryIndexResponse"
      },
      "RepositoryListResponse": {
        "properties": {
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/RepositoryResponse"
            },
            "type": "array",
            "title": "Repositories"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "repositories",
          "total"
        ],
        "title": "RepositoryListResponse"
      },
      "RepositoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "github_repo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "repo_name": {
            "type": "string",
            "title": "Repo Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "stars": {
            "type": "integer",
            "title": "Stars",
            "default": 0
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "indexing_status": {
            "$ref": "#/components/schemas/IndexingStatus"
          },
          "last_indexed_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed Commit"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "full_name",
          "repo_name",
          "default_branch",
          "is_private",
          "indexing_status"
        ],
        "title": "RepositoryResponse"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "repository_id": {
            "type": "integer",
            "title": "Repository Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "results",
          "repository_id",
          "query"
        ],
        "title": "SearchResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}